pyproject.toml
[project]
name = "polyai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "pyyaml>=6.0.3",
    "ruff>=0.15.8",
    "scipy>=1.17.1",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
]
